requirements.txt
redis
requests
json
sched
time
python-dotenv